splitting: make users private
