+++ src/lenskit/splitting/__init__.py
@@ -16,7 +16,7 @@ from ._holdout import HoldoutMethod, LastFrac, LastN, SampleFrac, SampleN
 from ._records import crossfold_records, sample_records
 from ._split import TTSplit
 from ._temporal import split_global_time, split_temporal_fraction
-from .users import crossfold_users, sample_users
+from ._users import crossfold_users, sample_users
 
 __all__ = [
     "TTSplit",
