+++ src/lenskit/splitting/_users.py
